[19/10/23] Bob can now die
Bob now has health and damage related functions
When health reaches zero, bob is frozen in place and teleported to the corner of the level.
Will have to update the coordinates if the level changes.
X value is the columns of the text file *50
Y value is the rows of the file * 50